Manual Refactor of endpoints to correct router post revase
